+++ webapp/src/components/modals/ObizOrderProcessModal.tsx
@@ -18,6 +18,7 @@ import { api } from "~/utils/api";
 import LayoutOrderStatus from "../obiz/LayoutOrderStatus";
 import { HiMiniShieldCheck } from "react-icons/hi2";
 import { formatter2Digits } from "~/utils/tools";
+import { useRouter } from "next/router";
 
 const ObizOfferVariableContent = ({
   step,
@@ -82,21 +83,30 @@ const ObizOfferVariableContent = ({
         );
       } else {
         if (!selectedArticles || !setSelectedArticles) return null;
+
+        const filteredArticles = articles
+          .filter((a) => a.kind === "fixed_price")
+          .sort((a, b) => {
+            if (a.publicPrice && b.publicPrice)
+              return a.publicPrice - b.publicPrice;
+            return 0;
+          });
+
         return (
           <>
-            <Box mt={10}>
+            <Box mt={10} mb={16}>
               <DiscountAmountBlock
                 kind="fixed_price"
                 amount={amount}
                 setAmount={setAmount}
-                articles={articles}
+                articles={filteredArticles}
                 selectedArticles={selectedArticles}
                 setSelectedArticles={setSelectedArticles}
               />
             </Box>
             <Button
               mt="auto"
-              mb={24}
+              mb={12}
               onClick={() => setStep("summary")}
               isDisabled={amount === 0}
               w="full"
@@ -167,6 +177,7 @@ export default function ObizOrderProcessModal(
   props: ObizOrderProcessModalProps
 ) {
   const { isOpen, onClose, offerId } = props;
+  const router = useRouter();
 
   const [amount, setAmount] = useState(0);
   const [step, setStep] = useState<Steps>("amount");
@@ -179,6 +190,7 @@ export default function ObizOrderProcessModal(
     onSuccess: ({ data: { payment_url } }) => {
       if (payment_url) window.location.href = payment_url;
     },
+    onError: () => router.push("/dashboard/order/error"),
   });
 
   const { data: offerResult } = api.offer.getById.useQuery({
@@ -197,8 +209,8 @@ export default function ObizOrderProcessModal(
   return (
     <Modal isOpen={isOpen} onClose={onClose} size="full">
       <ModalOverlay />
-      <ModalContent h="100dvh">
-        <ModalBody display="flex" flexDir="column" px={8} h="100dvh">
+      <ModalContent minH="full">
+        <ModalBody display="flex" flexDir="column" px={8} minH="full">
           {step !== "payment" && (
             <Box mt={8}>
               <BackButton